refactor: ProfileLikeRepository JPA에 의존하지 않게 DI 적용하여 구현체에서 선택하도록 변경 (#255)
