--- src/main/java/com/api/ttoklip/domain/profile/ProfileLikeRepository.java
@@ -1,6 +1,0 @@
-package com.api.ttoklip.domain.profile;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProfileLikeRepository extends JpaRepository<ProfileLike, Long>, ProfileLikeRepositoryCustom {
-}
