+++ src/main/java/com/api/ttoklip/domain/profile/ProfileLikeService.java
@@ -4,6 +4,8 @@ import static com.api.ttoklip.global.util.SecurityUtil.getCurrentMember;
 
 import com.api.ttoklip.domain.member.domain.Member;
 import com.api.ttoklip.domain.member.service.MemberService;
+import com.api.ttoklip.domain.profile.domain.ProfileLike;
+import com.api.ttoklip.domain.profile.domain.ProfileLikeRepository;
 import com.api.ttoklip.global.exception.ApiException;
 import com.api.ttoklip.global.exception.ErrorType;
 import com.api.ttoklip.global.success.Message;
